Move message specific hook and utils to relevant directories
