+++ src/modules/message/hooks/useVerifyMessageValidator.js
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import verifyMessageValidator from '../verifyMessageValidator';
+import verifyMessageValidator from '../utils/verifyMessageValidator';
 
 const useVerifyMessageValidator = (inputs) =>
   useMemo(() => verifyMessageValidator(inputs), [inputs]);
